divided huge controller into dedicated classes
