+++ src/main/java/pl/databucket/database/Bundle.java
@@ -1,4 +1,4 @@
-package pl.databucket.web.database;
+package pl.databucket.database;
 
 import java.util.Date;
