+++ src/main/java/pl/databucket/database/C.java
@@ -1,4 +1,4 @@
-package pl.databucket.web.database;
+package pl.databucket.database;
 
 public class C {
