+++ src/main/java/pl/databucket/database/Query.java
@@ -1,4 +1,4 @@
-package pl.databucket.web.database;
+package pl.databucket.database;
 
 import java.sql.Timestamp;
 import java.text.ParseException;
