+++ src/main/java/pl/databucket/database/TAB.java
@@ -1,4 +1,4 @@
-package pl.databucket.web.database;
+package pl.databucket.database;
 
 public class TAB {
 	public static final String GROUP = "_group";
