+++ src/main/java/pl/databucket/service/ResponseStatus.java
@@ -1,4 +1,4 @@
-package pl.databucket.web.service;
+package pl.databucket.service;
 
 public enum ResponseStatus {
 	OK, FAILED, NO_DATA;
