+++ src/main/java/pl/databucket/web/events/EventsLogController.java
@@ -1,0 +1,113 @@
+package pl.databucket.web.events;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import pl.databucket.database.C;
+import pl.databucket.database.Condition;
+import pl.databucket.database.FieldValidator;
+import pl.databucket.service.DatabucketService;
+import pl.databucket.service.ResponseBody;
+import pl.databucket.service.ResponseStatus;
+
+@CrossOrigin(origins = "*", allowedHeaders = "*")
+@RequestMapping("/api/events/log")
+@RestController
+public class EventsLogController {
+
+  private static final Logger logger = LoggerFactory.getLogger(EventsLogController.class);
+
+  private final DatabucketService databucketService;
+
+  public EventsLogController(DatabucketService databucketService) {
+    this.databucketService = databucketService;
+  }
+
+  @SuppressWarnings("unchecked")
+  @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<ResponseBody> getEventsLog(
+      @RequestParam(required = false) Optional<Integer> page,
+      @RequestParam(required = false) Optional<Integer> limit,
+      @RequestParam(required = false) Optional<String> sort,
+      @RequestParam(required = false) Optional<String> filter) {
+
+    ResponseBody rb = new ResponseBody();
+    try {
+      if (page.isPresent()) {
+        FieldValidator.mustBeGraterThen0("page", page.get());
+        rb.setPage(page.get());
+      }
+
+      if (limit.isPresent()) {
+        FieldValidator.mustBeGraterOrEqual0("limit", limit.get());
+        rb.setLimit(limit.get());
+      }
+
+      if (sort.isPresent()) {
+        FieldValidator.validateSort(sort.get());
+        rb.setSort(sort.get());
+      }
+
+      List<Condition> urlConditions = null;
+      if (filter.isPresent()) {
+        urlConditions = FieldValidator.validateFilter(filter.get());
+      }
+
+      Map<String, Object> result = databucketService.getEventsLog(page, limit, sort, urlConditions);
+
+      long total = (long) result.get(C.TOTAL);
+      rb.setTotal(total);
+
+      if (page.isPresent() && limit.isPresent()) {
+        rb.setTotalPages((int) Math.ceil(total / (float) limit.get()));
+      }
+
+      rb.setEventsLog((List<Map<String, Object>>) result.get(C.EVENTS_LOG));
+
+      return new ResponseEntity<>(rb, HttpStatus.OK);
+
+    } catch (Exception ee) {
+      return defaultException(rb, ee);
+    }
+  }
+
+  @DeleteMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+  public ResponseEntity<ResponseBody> clearEventsLog() {
+    ResponseBody rb = new ResponseBody();
+
+    try {
+      databucketService.clearEventsLog();
+      rb.setStatus(ResponseStatus.OK);
+      rb.setMessage("The events log has been cleaned.");
+      return new ResponseEntity<>(rb, HttpStatus.OK);
+    } catch (Exception ee) {
+      return defaultException(rb, ee);
+    }
+  }
+
+  private ResponseEntity<ResponseBody> customException(ResponseBody rb, Exception e, HttpStatus status) {
+    logger.warn(e.getMessage());
+    rb.setStatus(ResponseStatus.FAILED);
+    rb.setMessage(e.getMessage());
+    return new ResponseEntity<>(rb, status);
+  }
+
+  private ResponseEntity<ResponseBody> defaultException(ResponseBody rb, Exception e) {
+    logger.error("ERROR:", e);
+    rb.setStatus(ResponseStatus.FAILED);
+    rb.setMessage(e.getMessage());
+    return new ResponseEntity<>(rb, HttpStatus.INTERNAL_SERVER_ERROR);
+  }
+
+}
